Refactor with explorer UI
