+++ hollow-diff-ui/build.gradle
@@ -12,7 +12,7 @@ dependencies {
     api project(':hollow-ui-tools')
     implementation 'com.google.code.gson:gson:2.8.0'
 
-    compile 'org.eclipse.jetty:jetty-server:9.4.3.v20170317'
+    compileOnly 'org.eclipse.jetty:jetty-server:9.4.3.v20170317'
 
     testImplementation 'junit:junit:4.11'
 }
