+++ hollow-explorer-ui/src/main/java/com/netflix/hollow/explorer/ui/pages/BrowseSelectedTypePage.java
@@ -16,15 +16,15 @@
  */
 package com.netflix.hollow.explorer.ui.pages;
 
-import com.netflix.hollow.core.index.HollowPrimaryKeyIndex;
+import static com.netflix.hollow.tools.util.SearchUtils.getFieldPathIndexes;
+import static com.netflix.hollow.tools.util.SearchUtils.getOrdinalToDisplay;
+import static com.netflix.hollow.tools.util.SearchUtils.getPrimaryKey;
+import static com.netflix.hollow.tools.util.SearchUtils.parseKey;
+
 import com.netflix.hollow.core.index.key.PrimaryKey;
 import com.netflix.hollow.core.read.HollowReadFieldUtils;
 import com.netflix.hollow.core.read.engine.HollowTypeReadState;
-import com.netflix.hollow.core.read.engine.HollowTypeStateListener;
 import com.netflix.hollow.core.read.engine.object.HollowObjectTypeReadState;
-import com.netflix.hollow.core.schema.HollowObjectSchema;
-import com.netflix.hollow.core.schema.HollowSchema;
-import com.netflix.hollow.core.schema.HollowSchema.SchemaType;
 import com.netflix.hollow.explorer.ui.HollowExplorerUI;
 import com.netflix.hollow.explorer.ui.model.QueryResult;
 import com.netflix.hollow.explorer.ui.model.TypeKey;
@@ -81,7 +81,7 @@ public class BrowseSelectedTypePage extends HollowExplorerPage {
         }
         
         PrimaryKey primaryKey = getPrimaryKey(typeState.getSchema());
-        int fieldPathIndexes[][] = getFieldPathIndexes(primaryKey); 
+        int fieldPathIndexes[][] = getFieldPathIndexes(ui.getStateEngine(), primaryKey);
         
         List<TypeKey> keys = new ArrayList<>(pageSize);
         
@@ -94,13 +94,13 @@ public class BrowseSelectedTypePage extends HollowExplorerPage {
         String key = req.getParameter("key") == null ? "" : req.getParameter("key");
         Object parsedKey[] = null;
         try {
-            parsedKey = parseKey(primaryKey, key);
+            parsedKey = parseKey(ui.getStateEngine(), primaryKey, key);
         } catch(Exception e) {
             key = "";
         }
 
-        Integer ordinal =
-            getOrdinalToDisplay(req, key, parsedKey, selectedOrdinals, fieldPathIndexes);
+        HollowTypeReadState readTypeState = getTypeState(req);
+        Integer ordinal = getOrdinalToDisplay(ui.getStateEngine(), key, parsedKey, selectedOrdinals, fieldPathIndexes, readTypeState);
         if (ordinal != null && !ordinal.equals(-1) && "".equals(key)
                 && fieldPathIndexes != null) {
             // set key for the case where it was unset previously
@@ -118,38 +118,6 @@ public class BrowseSelectedTypePage extends HollowExplorerPage {
         ctx.put("key", key);
         ctx.put("ordinal", ordinal);
         ctx.put("display", displayFormat);
-    }
-
-    private Integer getOrdinalToDisplay(HttpServletRequest req, String key,
-            Object[] parsedKey, BitSet selectedOrdinals, int[][] fieldPathIndexes) {
-        HollowTypeReadState typeState = getTypeState(req);
-        int ordinal = req.getParameter("ordinal") == null ? -1
-            : Integer.parseInt(req.getParameter("ordinal"));
-        if ("".equals(key) && ordinal != -1) { // trust ordinal if key is empty
-            return ordinal;
-        } else if (!"".equals(key)) {
-            // verify ordinal key matches parsed key
-            if (ordinal != -1 && selectedOrdinals.get(ordinal)
-                    && recordKeyEquals(typeState, ordinal, parsedKey, fieldPathIndexes)) {
-                return ordinal;
-            } else {
-                HollowPrimaryKeyIndex idx = findPrimaryKeyIndex(typeState);
-                if (idx != null) {
-                    // N.B. - findOrdinal can return -1, the caller deals with it
-                    return findOrdinal(idx, key);
-                } else {
-                    // no index, scan through records
-                    ordinal = selectedOrdinals.nextSetBit(0);
-                    while (ordinal != -1) {
-                        if (recordKeyEquals(typeState, ordinal, parsedKey, fieldPathIndexes)) {
-                            return ordinal;
-                        }
-                        ordinal = selectedOrdinals.nextSetBit(ordinal + 1);
-                    }
-                }
-            }
-        }
-        return -1;
     }
 
     @Override
@@ -176,120 +144,28 @@ public class BrowseSelectedTypePage extends HollowExplorerPage {
     private TypeKey getKey(int recordIdx, HollowTypeReadState typeState, int ordinal, int[][] fieldPathIndexes) {
         if(fieldPathIndexes != null) {
             StringBuilder keyBuilder = new StringBuilder();
-            
+
             HollowObjectTypeReadState objState = (HollowObjectTypeReadState)typeState;
-            
+
             for(int i=0;i<fieldPathIndexes.length;i++) {
                 int curOrdinal = ordinal;
                 HollowObjectTypeReadState curState = objState;
-                
+
                 for(int j=0;j<fieldPathIndexes[i].length - 1;j++) {
                     curOrdinal = curState.readOrdinal(curOrdinal, fieldPathIndexes[i][j]);
                     curState = (HollowObjectTypeReadState) curState.getSchema().getReferencedTypeState(fieldPathIndexes[i][j]);
                 }
-                
+
                 if(i > 0)
                     keyBuilder.append(":");
-                
+
                 keyBuilder.append(HollowReadFieldUtils.fieldValueObject(curState, curOrdinal, fieldPathIndexes[i][fieldPathIndexes[i].length-1]));
             }
-            
+
             return new TypeKey(recordIdx, ordinal, keyBuilder.toString());
         }
-        
-        return new TypeKey(recordIdx, ordinal, "", "ORDINAL:"+ordinal);
-    }
-    
-    private PrimaryKey getPrimaryKey(HollowSchema schema) {
-        if(schema.getSchemaType() == SchemaType.OBJECT)
-            return ((HollowObjectSchema)schema).getPrimaryKey();
-        return null;
-    }
 
-    private int[][] getFieldPathIndexes(PrimaryKey primaryKey) {
-        if(primaryKey != null) {
-            int fieldPathIndexes[][] = new int[primaryKey.numFields()][];
-            for(int i=0;i<primaryKey.numFields();i++) {
-                fieldPathIndexes[i] = primaryKey.getFieldPathIndex(ui.getStateEngine(), i);
-            }
-            return fieldPathIndexes;
-        }
-        
-        return null;
-    }
-    
-    private HollowPrimaryKeyIndex findPrimaryKeyIndex(HollowTypeReadState typeState) {
-        if(getPrimaryKey(typeState.getSchema()) == null)
-            return null;
-        
-        for(HollowTypeStateListener listener : typeState.getListeners()) {
-            if(listener instanceof HollowPrimaryKeyIndex) {
-                if(((HollowPrimaryKeyIndex) listener).getPrimaryKey().equals(getPrimaryKey(typeState.getSchema())))
-                    return (HollowPrimaryKeyIndex)listener;
-            }
-        }
-        
-        return null;
-    }
-    
-    private int findOrdinal(HollowPrimaryKeyIndex idx, String keyString) {
-        Object[] key = parseKey(idx.getPrimaryKey(), keyString);
-        
-        return idx.getMatchingOrdinal(key);
-    }
-
-    private Object[] parseKey(PrimaryKey primaryKey, String keyString) {
-        // Split by the number of fields of the primary key
-        // This ensures correct extraction of an empty value for the last field
-        String fields[] = keyString.split(":", primaryKey.numFields());
-        
-        Object key[] = new Object[fields.length];
-        
-        for(int i=0;i<fields.length;i++) {
-            switch(primaryKey.getFieldType(ui.getStateEngine(), i)) {
-            case BOOLEAN:
-                key[i] = Boolean.parseBoolean(fields[i]);
-                break;
-            case STRING:
-                key[i] = fields[i];
-                break;
-            case INT:
-            case REFERENCE:
-                key[i] = Integer.parseInt(fields[i]);
-                break;
-            case LONG:
-                key[i] = Long.parseLong(fields[i]);
-                break;
-            case DOUBLE:
-                key[i] = Double.parseDouble(fields[i]);
-                break;
-            case FLOAT:
-                key[i] = Float.parseFloat(fields[i]);
-                break;
-            case BYTES:
-                key[i] = null; //TODO
-            }
-        }
-        return key;
-    }
-    
-    private boolean recordKeyEquals(HollowTypeReadState typeState, int ordinal, Object[] key, int[][] fieldPathIndexes) {
-        HollowObjectTypeReadState objState = (HollowObjectTypeReadState)typeState;
-        
-        for(int i=0;i<fieldPathIndexes.length;i++) {
-            int curOrdinal = ordinal;
-            HollowObjectTypeReadState curState = objState;
-            
-            for(int j=0;j<fieldPathIndexes[i].length - 1;j++) {
-                curOrdinal = curState.readOrdinal(curOrdinal, fieldPathIndexes[i][j]);
-                curState = (HollowObjectTypeReadState) curState.getSchema().getReferencedTypeState(fieldPathIndexes[i][j]);
-            }
-            
-            if(!HollowReadFieldUtils.fieldValueEquals(curState, curOrdinal, fieldPathIndexes[i][fieldPathIndexes[i].length-1], key[i]))
-                return false;
-        }
-        
-        return true;
+        return new TypeKey(recordIdx, ordinal, "", "ORDINAL:" + ordinal);
     }
 
     private HollowTypeReadState getTypeState(HttpServletRequest req) {
